+++ src/main/java/cool/furry/mc/forge/projectexpansion/config/ConfigMenu.java
@@ -87,8 +87,8 @@ public class ConfigMenu extends Screen {
             "gui.projectexpansion.config.collector_multiplier",
             1, 20, 1,
             __ -> (double) Config.collectorMultiplier.get(),
-            (__, newValue) -> Config.collectorMultiplier.set(newValue.longValue()),
-            (gs, option) -> new TranslationTextComponent("gui.projectexpansion.config.collector_multiplier").appendSibling(new StringTextComponent(String.format(": %s", option.get(gs))))
+            (__, newValue) -> Config.collectorMultiplier.set(newValue.intValue()),
+            (gs, option) -> new TranslationTextComponent("gui.projectexpansion.config.collector_multiplier").appendSibling(new StringTextComponent(String.format(": %s", (int) option.get(gs))))
         ));
 
         optionsRowList.addOption(new SliderPercentageOption(
@@ -96,39 +96,39 @@ public class ConfigMenu extends Screen {
             1, 20, 1,
             __ -> (double) Config.emcLinkItemLimitMultiplier.get(),
             (__, newValue) -> Config.emcLinkItemLimitMultiplier.set(newValue.intValue()),
-            (gs, option) -> new TranslationTextComponent("gui.projectexpansion.config.emc_link_item_limit_multiplier").appendSibling(new StringTextComponent(String.format(": %s", option.get(gs))))
+            (gs, option) -> new TranslationTextComponent("gui.projectexpansion.config.emc_link_item_limit_multiplier").appendSibling(new StringTextComponent(String.format(": %s", (int) option.get(gs))))
         ));
 
         optionsRowList.addOption(new SliderPercentageOption(
             "gui.projectexpansion.config.emc_link_emc_limit_multiplier",
             1, 20, 1,
             __ -> (double) Config.emcLinkEMCLimitMultiplier.get(),
-            (__, newValue) -> Config.emcLinkEMCLimitMultiplier.set(newValue.longValue()),
-            (gs, option) -> new TranslationTextComponent("gui.projectexpansion.config.emc_link_emc_limit_multiplier").appendSibling(new StringTextComponent(String.format(": %s", option.get(gs))))
+            (__, newValue) -> Config.emcLinkEMCLimitMultiplier.set(newValue.intValue()),
+            (gs, option) -> new TranslationTextComponent("gui.projectexpansion.config.emc_link_emc_limit_multiplier").appendSibling(new StringTextComponent(String.format(": %s", (int) option.get(gs))))
         ));
 
         optionsRowList.addOption(new SliderPercentageOption(
             "gui.projectexpansion.config.powerflower_multiplier",
             1, 20, 1,
             __ -> (double) Config.powerflowerMultiplier.get(),
-            (__, newValue) -> Config.powerflowerMultiplier.set(newValue.longValue()),
-            (gs, option) -> new TranslationTextComponent("gui.projectexpansion.config.powerflower_multiplier").appendSibling(new StringTextComponent(String.format(": %s", option.get(gs))))
+            (__, newValue) -> Config.powerflowerMultiplier.set(newValue.intValue()),
+            (gs, option) -> new TranslationTextComponent("gui.projectexpansion.config.powerflower_multiplier").appendSibling(new StringTextComponent(String.format(": %s", (int) option.get(gs))))
         ));
 
         optionsRowList.addOption(new SliderPercentageOption(
             "gui.projectexpansion.config.relay_bonus_multiplier",
             1, 20, 1,
             __ -> (double) Config.relayBonusMultiplier.get(),
-            (__, newValue) -> Config.relayBonusMultiplier.set(newValue.longValue()),
-            (gs, option) -> new TranslationTextComponent("gui.projectexpansion.config.relay_bonus_multiplier").appendSibling(new StringTextComponent(String.format(": %s", option.get(gs))))
+            (__, newValue) -> Config.relayBonusMultiplier.set(newValue.intValue()),
+            (gs, option) -> new TranslationTextComponent("gui.projectexpansion.config.relay_bonus_multiplier").appendSibling(new StringTextComponent(String.format(": %s", (int) option.get(gs))))
         ));
 
         optionsRowList.addOption(new SliderPercentageOption(
             "gui.projectexpansion.config.relay_transfer_multiplier",
             1, 20, 1,
             __ -> (double) Config.relayTransferMultiplier.get(),
-            (__, newValue) -> Config.relayTransferMultiplier.set(newValue.longValue()),
-            (gs, option) -> new TranslationTextComponent("gui.projectexpansion.config.relay_transfer_multiplier").appendSibling(new StringTextComponent(String.format(": %s", option.get(gs))))
+            (__, newValue) -> Config.relayTransferMultiplier.set(newValue.intValue()),
+            (gs, option) -> new TranslationTextComponent("gui.projectexpansion.config.relay_transfer_multiplier").appendSibling(new StringTextComponent(String.format(": %s", (int) option.get(gs))))
         ));
 
         addButton(new Button((width - BUTTON_WIDTH) / 2, height - DONE_BUTTON_TOP_OFFSET, BUTTON_WIDTH, BUTTON_HEIGHT, new TranslationTextComponent("gui.done"), (button) -> minecraft.displayGuiScreen(parentScreen)));
